Add i18n selector in mobile sidebar
