+++ src/components/Navbar/DesktopMenu.js
@@ -3,7 +3,6 @@ import Button from '@material-ui/core/Button';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import makeStyles from '@material-ui/core/styles/makeStyles';
 import IconButton from '@material-ui/core/IconButton';
-import TranslateIcon from '@material-ui/icons/Translate';
 import Favorite from '@material-ui/icons/Favorite';
 import ShoppingCart from '@material-ui/icons/ShoppingCart';
 import Badge from '@material-ui/core/Badge';
@@ -11,8 +10,7 @@ import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
 import LoginModal from '../Auth/LoginModal';
 import { useGoogleAuth } from '../Auth/GoogleAuthProvider';
-import { useLanguage } from '../../i18n/LanguageProvider';
-import { localList } from '../../i18n/locales';
+import LanguageSelector from '../Internationalization/LanguageSelector';
 
 const useStyles = makeStyles(theme => ({
   sectionDesktop: {
@@ -25,27 +23,13 @@ const useStyles = makeStyles(theme => ({
 
 function DesktopMenu() {
   const classes = useStyles();
-
-  const { setLocale: setLanguage, locale } = useLanguage();
   const { isInitialized, isSignedIn, signOut, googleUser } = useGoogleAuth();
-
   const [anchorEl, setAnchorEl] = useState(null);
 
-  const [languageSelectAnchorEl, setLanguageSelectAnchorEl] = useState(null);
-
   const isMenuOpen = Boolean(anchorEl);
-  const isLanguageMenuOpen = Boolean(languageSelectAnchorEl);
 
   const handleProfileMenuOpen = event => {
     setAnchorEl(event.currentTarget);
-  };
-
-  const handleLanguageMenuOpen = event => {
-    setLanguageSelectAnchorEl(event.currentTarget);
-  };
-
-  const handleLanguageMenuClose = () => {
-    setLanguageSelectAnchorEl(null);
   };
 
   const handleMenuClose = () => {
@@ -55,11 +39,6 @@ function DesktopMenu() {
   const handleLogout = () => {
     signOut();
     handleMenuClose();
-  };
-
-  const selectLanguage = local => {
-    setLanguage(local);
-    handleLanguageMenuClose();
   };
 
   const menuId = 'primary-search-account-menu';
@@ -77,27 +56,6 @@ function DesktopMenu() {
     </Menu>
   );
 
-  const languageMenuId = 'language-menu';
-  const renderLanguageMenu = (
-    <Menu
-      anchorEl={languageSelectAnchorEl}
-      anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
-      id={languageMenuId}
-      keepMounted
-      transformOrigin={{ vertical: 'top', horizontal: 'right' }}
-      open={isLanguageMenuOpen}
-      onClose={handleLanguageMenuClose}
-    >
-      {localList.map(item => {
-        return (
-          <MenuItem key={Math.random()} onClick={() => selectLanguage(item)}>
-            {item.languageName}
-          </MenuItem>
-        );
-      })}
-    </Menu>
-  );
-
   return (
     <div className={classes.sectionDesktop}>
       {!isInitialized || !isSignedIn ? <LoginModal loading={isInitialized} /> : null}
@@ -112,16 +70,7 @@ function DesktopMenu() {
         </>
       ) : null}
 
-      <Button
-        aria-label="show languages"
-        aria-controls={languageMenuId}
-        aria-haspopup="true"
-        onClick={handleLanguageMenuOpen}
-        startIcon={<TranslateIcon />}
-        endIcon={<ExpandMoreIcon />}
-      >
-        {locale.languageName}
-      </Button>
+      <LanguageSelector />
 
       <IconButton aria-label="show 4 new mails" color="inherit">
         <Badge badgeContent={4} color="primary">
@@ -130,7 +79,6 @@ function DesktopMenu() {
       </IconButton>
 
       {renderMenu}
-      {renderLanguageMenu}
     </div>
   );
 }
