+++ src/components/Sidebar/Sidebar.js
@@ -1,7 +1,12 @@
 import React, { useState } from 'react';
-import { AccountCircle, Inbox, Mail, MoreVert } from '@material-ui/icons';
+import AccountCircle from '@material-ui/icons/AccountCircle';
+import Inbox from '@material-ui/icons/Inbox';
+import Mail from '@material-ui/icons/Mail';
+import Button from '@material-ui/core/Button';
+
 import IconButton from '@material-ui/core/IconButton';
 import MenuIcon from '@material-ui/icons/Menu';
+import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import {
   Divider,
   Drawer,
@@ -16,6 +21,7 @@ import useStyles from './SidebarStyles';
 import logo from '../../assets/app-logo.png';
 import { useGoogleAuth } from '../Auth/GoogleAuthProvider';
 import LoginModal from '../Auth/LoginModal';
+import LanguageSelector from '../Internationalization/LanguageSelector';
 
 function Sidebar() {
   const classes = useStyles();
@@ -43,21 +49,38 @@ function Sidebar() {
         <div className={classes.sidebarLogo} style={{ backgroundImage: `url(${logo})` }} />
       </div>
       <Divider />
-      {googleUser && googleUser.profileObj ? (
-        <>
-          <div className={classes.username}>
-            <AccountCircle className={classes.accountCircle} />
-            <h3 className={classes.sidebarUser}>Hi, {googleUser.profileObj.name}</h3>
-            <Hidden smUp>
-              <div>
-                <IconButton
+      <h2 className={classes.menuHeader}>Categories</h2>
+      <List>
+        {['Indian wear', 'Western wear', 'Jewellery', 'Accessories'].map((text, index) => (
+          <ListItem button key={text} className={classes.listItemStyle}>
+            {index % 2 === 0 ? (
+              <Inbox style={{ marginRight: '15px' }} />
+            ) : (
+              <Mail style={{ marginRight: '15px' }} />
+            )}
+            <ListItemText primary={text} />
+          </ListItem>
+        ))}
+      </List>
+      <Hidden smUp>
+        <div className={classes.mobileMenuContainer}>
+          <Divider />
+          <LanguageSelector style={{ width: '100%', justifyContent: 'space-evenly' }} />
+          {googleUser && googleUser.profileObj ? (
+            <>
+              <div className={classes.username}>
+                <Button
                   aria-label="more"
                   aria-controls="long-menu"
                   aria-haspopup="true"
                   onClick={onDotMenuClick}
+                  startIcon={<AccountCircle className={classes.accountCircle} />}
+                  endIcon={<ExpandMoreIcon />}
+                  style={{ width: '100%', justifyContent: 'space-evenly' }}
                 >
-                  <MoreVert />
-                </IconButton>
+                  {googleUser.profileObj.name}
+                </Button>
+
                 <Menu
                   id="long-menu"
                   anchorEl={anchorEl}
@@ -74,31 +97,16 @@ function Sidebar() {
                   <MenuItem onClick={onLogoutClick}>Logout</MenuItem>
                 </Menu>
               </div>
-            </Hidden>
-          </div>
-          <Divider />
-        </>
-      ) : (
-        <Hidden smUp>
-          <div className={classes.login}>
-            {!isInitialized || !isSignedIn ? <LoginModal loading={isInitialized} /> : null}
-          </div>
-          <Divider />
-        </Hidden>
-      )}
-      <h2 className={classes.menuHeader}>Categories</h2>
-      <List>
-        {['Indian wear', 'Western wear', 'Jewellery', 'Accessories'].map((text, index) => (
-          <ListItem button key={text} className={classes.listItemStyle}>
-            {index % 2 === 0 ? (
-              <Inbox style={{ marginRight: '15px' }} />
-            ) : (
-              <Mail style={{ marginRight: '15px' }} />
-            )}
-            <ListItemText primary={text} />
-          </ListItem>
-        ))}
-      </List>
+            </>
+          ) : (
+            <>
+              <div className={classes.login}>
+                {!isInitialized || !isSignedIn ? <LoginModal loading={isInitialized} /> : null}
+              </div>
+            </>
+          )}
+        </div>
+      </Hidden>
     </div>
   );
